refactor: Lotto패키지 model로 변경
